+++ homeassistant/components/nest/climate_sdm.py
@@ -4,6 +4,7 @@ from typing import Optional
 
 from google_nest_sdm.device import Device
 from google_nest_sdm.device_traits import FanTrait, TemperatureTrait
+from google_nest_sdm.event import AsyncEventCallback, EventMessage
 from google_nest_sdm.exceptions import GoogleNestException
 from google_nest_sdm.thermostat_traits import (
     ThermostatEcoTrait,
@@ -36,10 +37,9 @@ from homeassistant.components.climate.const import (
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import ATTR_TEMPERATURE, TEMP_CELSIUS
 from homeassistant.exceptions import PlatformNotReady
-from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.typing import HomeAssistantType
 
-from .const import DATA_SUBSCRIBER, DOMAIN, SIGNAL_NEST_UPDATE
+from .const import DATA_SUBSCRIBER, DOMAIN
 from .device_info import DeviceInfo
 
 # Mapping for sdm.devices.traits.ThermostatMode mode field
@@ -94,7 +94,7 @@ async def async_setup_sdm_entry(
     async_add_entities(entities)
 
 
-class ThermostatEntity(ClimateEntity):
+class ThermostatEntity(ClimateEntity, AsyncEventCallback):
     """A nest thermostat climate entity."""
 
     def __init__(self, device: Device):
@@ -126,17 +126,12 @@ class ThermostatEntity(ClimateEntity):
 
     async def async_added_to_hass(self):
         """Run when entity is added to register update signal handler."""
-        # Event messages trigger the SIGNAL_NEST_UPDATE, which is intercepted
-        # here to re-fresh the signals from _device.  Unregister this callback
-        # when the entity is removed.
         self._supported_features = self._get_supported_features()
-        self.async_on_remove(
-            async_dispatcher_connect(
-                self.hass,
-                SIGNAL_NEST_UPDATE,
-                self.async_write_ha_state,
-            )
-        )
+        self.async_on_remove(self._device.add_event_callback(self))
+
+    async def async_handle_event(self, event_message: EventMessage):
+        """Let Home Assistant know device state has been updated."""
+        self.async_write_ha_state()
 
     @property
     def temperature_unit(self):
